+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -130,6 +130,12 @@ Improvements to clang-tidy
   <clang-tidy/checks/modernize-use-override>` now supports `OverrideSpelling`
   and `FinalSpelling` options.
 
+- New :doc:`openmp-exception-escape
+  <clang-tidy/checks/openmp-exception-escape>` check.
+
+  Analyzes OpenMP Structured Blocks and checks that no exception escapes
+  out of the Structured Block it was thrown in.
+
 - New :doc:`openmp-use-default-none
   <clang-tidy/checks/openmp-use-default-none>` check.